Clearing up default exclusion protocols and cipher-suites setup by 9.4.8 version of jetty as they could potentially break some setups
Wrapped this around a system environment flag to switch the behavior to defaults set by jetty in future
Also removed the code to fall back to weak ciphers since no one has ever seen an issue with the default configurable list
