+++ jetty9/src/main/java/com/thoughtworks/go/server/util/GoSslSocketConnector.java
@@ -1,5 +1,5 @@
 /*
- * Copyright 2017 ThoughtWorks, Inc.
+ * Copyright 2018 ThoughtWorks, Inc.
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
@@ -62,6 +62,10 @@ public class GoSslSocketConnector implements GoSocketConnector {
         httpsConfig.addCustomizer(new ForwardedRequestCustomizer());
 
         SslContextFactory sslContextFactory = new SslContextFactory();
+        if(systemEnvironment.get(SystemEnvironment.GO_SSL_CONFIG_CLEAR_JETTY_DEFAULT_EXCLUSIONS)){
+            sslContextFactory.setExcludeProtocols();
+            sslContextFactory.setExcludeCipherSuites();
+        }
         sslContextFactory.setKeyStorePath(keystore.getPath());
         sslContextFactory.setKeyStorePassword(password);
         sslContextFactory.setKeyManagerPassword(password);
