+++ jetty9/src/test/java/com/thoughtworks/go/server/Jetty9ServerTest.java
@@ -97,12 +97,12 @@ public class Jetty9ServerTest {
         when(systemEnvironment.get(SystemEnvironment.RESPONSE_BUFFER_SIZE)).thenReturn(1000);
         when(systemEnvironment.get(SystemEnvironment.IDLE_TIMEOUT)).thenReturn(2000);
         when(systemEnvironment.configDir()).thenReturn(configDir = temporaryFolder.newFile());
-        when(systemEnvironment.get(SystemEnvironment.GO_SSL_CONFIG_ALLOW)).thenReturn(true);
         when(systemEnvironment.get(SystemEnvironment.GO_SSL_RENEGOTIATION_ALLOWED)).thenReturn(true);
         when(systemEnvironment.getJettyConfigFile()).thenReturn(new File("foo"));
         when(systemEnvironment.isSessionCookieSecure()).thenReturn(false);
         when(systemEnvironment.sessionTimeoutInSeconds()).thenReturn(1234);
         when(systemEnvironment.sessionCookieMaxAgeInSeconds()).thenReturn(5678);
+        when(systemEnvironment.get(SystemEnvironment.GO_SSL_CONFIG_CLEAR_JETTY_DEFAULT_EXCLUSIONS)).thenReturn(true);
 
 
         when(sslSocketFactory.getSupportedCipherSuites()).thenReturn(new String[]{});
